[launcher] Change enabling systemd service to link
Currently installed services are enabled by default and started by systemd on
system restart. In this commit services are linked to systemd and not started
automatically. launcher.startAllServices starts installed services on launcher
create. This functional change is required by adding users implementation. We
should not start services till we get current users.

For service cashing we do not need to delete service on remove but just stop.
This change is preparing for above implementation.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>