[servicemanager] Fix service update from cloud
Current update implementation is buggy and leads to runtime error: index
out of range in case of multiple services are updated. Changing array size
during of iteration on it leads to this kind of errors. Fix is to use
a map with service id as key. This map keeps info about installed services
and new services received from cloud. First, the map is filled by
installed services info then updated with new info received from cloud.
Second, map is analyzed as following:
- if map item contains installed and new service infos then the service
  should be updated (if new service has higher version);
- if map item contains only new service info then the service should be
  installed;
- if map item contains only installed service info then the service should
  be deleted.

Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>