[launcher] Remove service from watch systemd list on service update
There is no way to recognize if service was restart due to update or due to
some external reason such a fatal error etc. As result during update status
of service switched to error status. Currently it affects nothing as
service status is not send to cloud. But in general this behaviour is not
correct. The solution is to remove service from service state changed list
on service update and update service status manually.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>