[storagemanger] Fix few services installing
Fix of issue that  doesn't allow to install more than one services at once.
In processAmqpMessage function when servicesMap is filled, it uses pointers
of value created in "for range" loop. But this value is created only once
and reused on each iteration. As result map contains pointers to last
value. The fix is to create local variable, assign value from "for range"
and fill map with pointer to this local variable.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>
Reviewed-by: Oleh Mykytiuk <oleh_mykytiuk@epam.com>