[launcher] Return error in removeService without interrupting the sequence
During service removal, we do not interrupt procedure on first error but
try to perform all required steps. The patch is store err in local
variable, performs all steps and returns error if it was set. In this case
we perform all removing steps and notifies the initiator about the error.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>
Reviewed-by: Oleh Mykytiuk <oleh_mykytiuk@epam.com>