[launcher] Set disk quota as sum of storage and state sizes
We can set common disk quota for storage and state. Currently we treat
state as part of storage. But this is not obvious that setting state size
the user should set storage size as well. And besides storage size should
more or equal state size.
We agreed to set disk quota as sum of storage and state sizes. In this
case, if user doesn't want to use storage and sets only state size, the
quota will be applied as state size.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>
Reviewed-by: Oleh Mykytiuk <oleh_mykytiuk@epam.com>