[alerts] Set journal cursor to tail at the initialization
On the target, journal is stored in RAM only. It is starts from the
beginning after reboot. We can't restore the cursor. According to the doc,
if there is no cursor in the journal, it will be sought to closest by time
entry. But it causes getting cursor error. Setting cursor to tail at the
beginning resolves the issue.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>
Reviewed-by: Oleh Mykytiuk <oleh_mykytiuk@epam.com>