[NUNCOTA-13] Create vis identification module
For nuance project we require another (than vis) way to retrieve system id
and user claims. For this purpose, we've implemented identification
modules, which can be selected and configured in the config file. As we
have a lot of problem with golang plugins, the modules are implemented as
part of SM and selection is done by using build tags (with_vis_plugin).
This commit contains common for all modules package and VIS module.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>