[networkmanager] Save vlan interface name in DB
At service startup, we read from the database to check
if there is a network for the provider. If it exists,
we verify the existence of VLAN and bridge interfaces.
If they already exist, we do not create them; otherwise,
we do. The bridge interface name is created based on the
service provider's ID, while the VLAN name is generated randomly.
Therefore, at startup, the VLAN interface does not exist because
a new name has been generated. However, we store the VLAN ID in the
database. When we set up the VLAN ID for the newly created VLAN, we
encounter an error indicating that a VLAN interface with that ID
already exists. Hence, we need to store not only the VLAN ID but
also the VLAN interface name in the database to prevent this error.

Signed-off-by: Mykola Solianko <mykola_solianko@epam.com>